requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
ordereddict==1.1
python-slugify==4.0.1
requests==2.25.0
six==1.15.0
text-unidecode==1.3
urllib3==1.26.2
WooCommerce==2.1.1
